ml_service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
torch==2.4.1
torchvision==0.19.1
transformers==4.44.2
Pillow==10.4.0
python-multipart==0.0.9
google-generativeai==0.8.1
python-dotenv==1.0.1
